refactor(gamesimulator): split GameSimulator into cohesive helpers
Break the 1317-line GameSimulator orchestrator into package-private
helpers so each file owns a single concern and the main engine stays a
thin dispatcher over the snap loop.

New helpers (all <210 lines): PlayEventFactory, PenaltyEmitter,
InjuryEmitter, ScoringSequencer, SpecialTeams, EndOfHalfPlays,
PeriodController, DownProgression. GameSimulator drops to 540 lines
(most of which is the five overloaded convenience constructors).

Behavior is preserved — all 363 tests pass without seed changes. One
test reference updated: GameSimulator.concludeOvertimePossession →
PeriodController.concludeOvertimePossession.
